+++ packages/demo/config/load-links.js
@@ -27,6 +27,7 @@ const rawLinks = [
   { label: 'protractor', path: '/tools/protractor' },
   { label: 'rotatable', path: '/tools/rotatable' },
   { label: 'drag', path: '/drag' },
+  { label: 'pie-toolbox', path: '/pie-toolbox' },
 ];
 
 module.exports.loadLinks = () => {