+++ packages/pie-toolbox/src/feedback.js
@@ -1,0 +1,11 @@
+import {
+  getFeedbackForCorrectness,
+  getFeedback,
+  defaults
+} from "./code/feedback";
+
+export {
+  getFeedbackForCorrectness,
+  getFeedback,
+  defaults
+};