+++ packages/pie-toolbox/src/icons.js
@@ -1,0 +1,21 @@
+import {
+    Correct,
+    CorrectResponse,
+    Incorrect,
+    Instructions,
+    LearnMore,
+    NothingSubmitted,
+    PartiallyCorrect,
+    ShowRationale,
+} from './code/icons';
+
+export {
+    Correct,
+    CorrectResponse,
+    Incorrect,
+    Instructions,
+    LearnMore,
+    NothingSubmitted,
+    PartiallyCorrect,
+    ShowRationale,
+};