+++ packages/pie-toolbox/src/math-evaluator.js
@@ -1,0 +1,3 @@
+import areValuesEqual from './code/math-evaluator';
+
+export default areValuesEqual;